+++ CHANGELOG.md
@@ -11,6 +11,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - New DfE font and header have been applied to the application.
 - The links in the footer have been updated to the latest versions.
 
+### Fixed
+
+- The primary contact for school in the export no longer uses the first contact
+  if one is not set.
+- The primary contact for the school is now labelled in the export.
+- The primary contact for the incoming trust in the export no longer uses the
+  first contact if one is not set.
+- The primary contact for the incoming trust is now labelled in the export.
+- The primary contact for the outgoing trust in the export no longer uses the
+  first contact if one is not set.
+- The primary contact for the outgoing trust is now labelled in the export.
+
 ## [Release-91][release-91]
 
 ### Added
@@ -60,6 +72,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Corrected the Edit link for Sharepoint links on Conversions, it was pointing
   to the Transfers edit page
+
+### Added
+
+- The headteacher contact selected in the 'Confirm the headteacher’s details'
+  task is now included in the RPA, SUG and FA letters export. The contacts role
+  will always be exported as 'Headteacher'.
+- The incoming trust CEO contact selected in the 'Confirm the incoming trust
+  CEO’s details' task is now included in the RPA, SUG and FA letters export. The
+  contacts role will always be exported as 'CEO'.
 
 ## [Release-88][release-88]